refactor(derive): Clarify parse is really about attributes
