+++ clap_derive/src/lib.rs
@@ -22,10 +22,10 @@ use proc_macro::TokenStream;
 use proc_macro_error::proc_macro_error;
 use syn::{parse_macro_input, DeriveInput};
 
+mod attr;
 mod derives;
 mod dummies;
 mod item;
-mod parse;
 mod utils;
 
 /// Generates the `ValueEnum` impl.
